requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
pymupdf
langchain
chromadb
openai
python-dotenv
jinja2
selenium
